Make it easier and less error prone to use rotating-secrets
Although ideally all components would have a `Supplier[SecretState]`, we
can't change Play and the verification it does on the
`play.http.secret.key` (which only models allowing one key, not several).
Grabbing the active secret from the supplier and passing it to the
configuration means that it'll get picked up by the quality checks in
`play.api.http.HttpConfiguration#getSecretConfiguration` - so we don't
need to set a dummy value.