Various tweaks, getting rid of AWS Secret Manager support, adding Lambda
Regarding AWS Secrets Manager, see it's secret-rotation lifecycle:

https://docs.aws.amazon.com/secretsmanager/latest/userguide/rotating-secrets-lambda-function-overview.html

The `createSecret`, `setSecret`, `testSecret`, `finishSecret` lifecycle
that Secrets Manager uses to update a secret is an unhelpful complication
for the use case of rotating the Play application secret, where there's no
meaningful, useful & easy test for a secret. The Secrets Manager API also
only exposes the `CreatedDate` of a secret, not when it was finalised,
which doesn't marry well with `play-secret-rotation's need for a
'publication time' of a secret, from which a usage-delay will determine
when that secret becomes active.