Reduce logging warning about accepting old secrets
Following on from https://github.com/guardian/play-secret-rotation/pull/2,
looking at Ophan's logs for the "Accepted decode with non-active key"
warning message, we can see that in the recent 24 hour period, there
are still 177 logged warnings:

https://logs.gutools.co.uk/s/ophan/goto/19c2ea70-f1f5-11ee-be1c-d5e1a37b2c42

PR #2 was an update that meant encountering an old key would only be
logged as a warning if we were in the final third (1/3) of the overlap period
for the transition - usually the overlap period is 2 hours, so we'd only log a
warning if we received something signed with an old secret in the last 40
minutes of the overlap period.

To reduce the level of logged messages further, we now restrict ourselves
to logging a warning in the final 1/5 of the overlap period - 24 minutes.
Based on Ophan logging, this won't get rid of warnings altogether, but will
reduce them quite sharply.