Include scalatest test results on stdout
By default, using scalatest’s -u flag (which outputs test results to a
junit xml file) stops the results from appearing on stdout: there’s
still a final summary, but not the test-by-test output. This can make it
a bit difficult to tell at a glance which tests have failed, so I’m
adding the -o flag to restore the stdout output.